Модуляция (сделал чуть больше: разметку будущего приложения)
