+++ js/config.js
@@ -1,0 +1,7 @@
+import {generateTestData} from './dataGenerator.js';
+
+const config = {
+  getData: generateTestData,
+};
+
+export {config};
